requirements.txt
Flask>=2.0
python-dotenv>=0.19
twilio>=7.0
pytest>=7.0
gunicorn>=20.1.0
